perf(ssr): remove `filterProperties` (#5101)
