+++ packages/@lwc/ssr-compiler/src/compile-js/generate-markup.ts
@@ -31,11 +31,6 @@ const bGenerateMarkup = esTemplate`
         tagName = tagName ?? ${/*component tag name*/ is.literal};
         attrs = attrs ?? Object.create(null);
         props = props ?? Object.create(null);
-        props = __filterProperties(
-            props,
-            publicFields,
-            privateFields,
-        );
         const instance = new ${/* Component class */ is.identifier}({
             tagName: tagName.toUpperCase(),
         });
@@ -43,7 +38,7 @@ const bGenerateMarkup = esTemplate`
         __establishContextfulRelationship(contextfulParent, instance);
         ${/*connect wire*/ is.statement}
 
-        instance[__SYMBOL__SET_INTERNALS](props, attrs);
+        instance[__SYMBOL__SET_INTERNALS](props, attrs, publicFields, privateFields);
         instance.isConnected = true;
         if (instance.connectedCallback) {
             __mutationTracker.enable(instance);
@@ -128,7 +123,6 @@ export function addGenerateMarkupFunction(
     program.body.unshift(
         bImportDeclaration({
             fallbackTmpl: '__fallbackTmpl',
-            filterProperties: '__filterProperties',
             hasScopedStaticStylesheets: undefined,
             mutationTracker: '__mutationTracker',
             renderAttrs: '__renderAttrs',
